Fix: div overflow. Feature: Selecting an item in the "SearchSection" will highlight the item in the "GlobeSection" and vice versa.
That took several hours to figure out. It turned out to be simple. I don't know what my error was at first, but I kept digging to figure it out and didn't notice how simple it really was for a while. Maybe it was that my head was having a hard time wrapping around it. Gah.