Fix: Text alignment for paragraph when words wrap.
The paragraph element was left-aligned due to the containing div's className "items-start", but if the words wrapped, they were horizontally centered. I made them left-aligned by setting the className "text-left" to the <p> object that contained each entry. This also made adjacent entries blend together, so I gave each <p> a border and a small margin too. And rounded it to look nice.