Feature: Globe now rendering points of interest.
Still having a problem with calculating normalized mouse position (e.clientX != 0 on the far left of the canvas), and I had trouble getting the POI info popup element to follow the mouse. I tried copy-pasting my code from the Chris Courses tutorials, but for some reason, copy-pasting the same barebones code resulted in an error from gsap on this program, saying that it couldn't set non-existent properties "x" and "y". They didn't exist when I was doing the tutorials either, so why is it complaining now?