Refactor: Updated the Scene to use the "selected POI" state slice.
And along the way I discovered that setting up gsap.to)...) in a useEffect(...) would cause that familiar brief flicker when selecting a new search result (or a permanent dark color on all POIs if I _de_-selected a POI search result). Was that the culprit this whole time? _Why_ is that a culprit?