Fix: Crash on selecting POI after new search + flickering on selection, . Refactor: Simplified selection.
Re-combined the states of "selectedPoi" (along with "prevSelectedPoi") with "pointsOfInterest" in a single Redux reducer so that running a new search will reset the selected POI objects and not cause the "de-highlight old selection" logic to blow up due to a non-existent object. I also fixed the flickering on selecting a POI (it was GSAP again; need to figure out what to do about that). And I found that I didn't need to fuss about in the Scene with material changes immediately upon selecting a new POI because the new useEffect() hook that depends on changes in selectedPoi will handle all that for me.