Experiment: Using the "Delaunator" library to create a delauney vertex mesh out of an unordered set of vertices.
I think that I'll be needing to switch to a 3D library. The 2D one is fast and works great (https://github.com/mapbox/delaunator), but at the edge of the latitudes and longitudes it wraps around. So I think that I'm going to need a 3D version of this library (https://github.com/d3/d3-delaunay, though this one is not as well documented for getting started).