Experiment: Successfully used the "d3-geo-voronoi" to create a half-sphere mesh out of lat-long points.
Note: The library uses the "long-lat" convention rather that "lat-long". I'll have to separate the use of the "delaunator" library for when I want to do regions for non-sphere maps. This "d3-geo-voronoi" library should (crossing fingers) work for the regions that I want to build out of an arbitrary set of lat-long coordinates.