Experiment: Cleaned up the "bi-spherical" interpolation between 3 vectors on the surface of a sphere.
Renaming and simplifying. This follows the same logic as the flat vector version, filling out a triangle with evenly spaced points, but this time a triangle on the surface of a sphere.

I did some profiling using the browser dev tools' "Performance" tab and found that, for a large triangle, it can take 2-3ms. Cool.

In comparison, the simple vector-based interpolation was so fast that it did not even appear in the profile (yes, I was searching for it). I had to add an artificial delay by making a loop that would print to the console, and just 10 loops writing to the console, the function appeared for only 1ms. I removed the console-printing loop, and the function disappeared from the Performance profiling entirely.

So yeah. The vector-based bilinear interpolation + rescaling, while not as precise as spherical interpolation and a bit uglier, is much more performant, and as long as it gives a good enough polygon...great!