Experiment: Re-worked the logic for creating points around a POI so that they are evenly space regardless of lat/long.
Ended up figuring out how to use ThreeJs to rotate a vector. I'm not sure what I was doing wrong, but I've been hacking at this thing for at least a couple hours and it wasn't working, but now it is. I'll take it.

I think that I'm going to need to use this because these "mini points" to create the region "vertices shell" because this is only way for me to have control over the ordering of the outer-most region vertices, and I need that control over the ordering so that I can use the "ear clipping" algorithm to triangulate the region vertices (no, delaunay won't work; it was build to form 3D solids, not shells).

Still need to clean up though and add more functionality to these "mini points" and restrain a region to single "POI" surrounded by adjsutable mini points.