Fix: Lack of error handling in ConvertLatLongToXYZ and ConvertXYZToLatLong.
As a result, I spent over 2 hours trying to figure out the problem only to find that I had simply forgotten to include a necessary variable. I wish that Javascript could be more strict about not allowing functions to be called unless all arguments are present. Why should I have to manually add error handling just to check for that?