Experiment: W00t! I'm able to click and drag a pin.
I don't think well. I diagnose a lot, but I can't think. This is my recent conclusion about myself.

This has taken me over 12hrs to figure out today. I spent a few hours working on state machine refactoing and starting to move away from the "whereLatLongArray" in the "editState" and towards a single "where pin" (for now), and then I lost a few hours to diagnosing errors that turned out to be missing arguments, lost several hours trying to figure out how to get a newly created PinMesh from the EditRegion to be added to the collection of interactable meshes in the Scene (turned out that I wasn't setting a state flag in EditRegion at the right time), and I lost at least a couple hours to a logic fault in my recursion.