Experiment: Trying to get region to redraw when the pin moves, but having trouble.
Ideally, I want to update the Edit state's regionBoundaries collection, but the collection seems to be (at least from the "map(...)" function) immutable, which means that, if I want to update it every frame, I have to find the pin being moved and replace it. This is taking time to figure out how to work myself out of the architectural pit that I've gotten myself into.