Fix: Mouse click not disabling, meshes not updating for intersection calculations.
In moving from the "mousInfoRef" to the mouse state machine, I forgot to implement a way to turn off the mouse click. Did that. Then I broke the "update mesh info for intersection raycaster" calculations by moving towards triggering on regionBoundary changes, but I fogot to change the useEffect(...) dependency.