Fix: Got the region resizing with my pin moves. Gah.
The problem was that the region calculations were based on the original region boundaries, an object stuck in time, even though the useEffect surrounding the region calculations was set to trigger when the editState.regionBoundariesChanged. So I changed the calculations to be entirely dependent on the editState.RegionBoundaries, and now it works. It was that this whole time.

And it looks like my "ear clipping" algorithm works! Mostly! It crashes though when a point is moved to leapfrog another point. Gotta figure that out.