Refactor (in-progress): Trying to move mouse click handling to its own file
But also trying to fix a race condition that's been irritating me for a while. It seems like I always have to click on the globe twice to generate a region. I thought that running reduxDispatch(...) would _schedule_ a state change, but apparently it does it immediately, which presents a problem when I say that "mouse is up" followed by a call to "disable mouse click" in the same frame. Need to re-think this.