Refactor: Starting to incorporate a quaternion rotator for moving the pins.
Not moving the pins yet (almost though). Once I figured out what info to store (and how I wanted to do it), the updated design came together pretty quickly. I will (1) store the normalized world pos of the cursor on the "mouse down" event, (2) create a quaternion rotor whenever the mouse was clicked on a moveable pin and the mouse is currently moving, and (3) apply the rotor to the pin's world pos. I've done (1) and (2), but still need to think through (3).