Refactor: Moving pins again + fixed "snap to cursor" problem when clicking and draggin
Quaternions were my answer. When starting the click-and-drag routine, I have get the intersection between the cursor and mesh + the intersection between the cursor and the globe behind the mesh. Then I create an offset quaternion based on the difference between those two points, and carry that on through the rest of the click-and-drag routine. It feels right now.