Investigate: Why is the region mesh semi-disappearing from the raycasting?
When I place a region, the raycast intersection works as expected. When I move the region, any part of the mesh that is beyond the original mesh stops intersecting, but any part that is still within the original boundaries works. When I move the mesh outside the original boundaries and then move the cursor over where it used to be, the intersection can't find it, even though I can confirm that the mesh is still there. If I move the mesh back into the original region, the intersection works again.

Looks like the intersection is working on two different sets of geometry: the current mesh geometry, and the original mesh geometry when it was first placed.

No, the mesh isn't disappearing form the children in the scene. I tried making sure that the intersections happened with the current scene children on every frame (no filtering to improve performance).

Yes, the geometry is changing when I move the mesh. I confirmed it with prior debugging.

What is wrong with this thing?!