Fix: Region move + geometry deformation
Had to update the bounding sphere after updating the geometry:
'regionMeshRef.current.geometry.computeBoundingSphere()'

See solution here:
https://discourse.threejs.org/t/raycaster-dont-intersect-with-geometry-points-after-updating-the-position-vertices/37858

But I can't rely on mesh movement via quaternions during click-and-drag because the quaternion rotation is a separate field from the verticies, which are calculated in-place on the globe. When I tried both re-creating the geometry when boundary markers are deformed + quaternions for rotating the mesh, then the re-calculated vertices had the quaternion applied after they were created, which rotated them way outside of where they were supposed to be.

So don't use transforms on deformed meshes (unless you find a magic algorithm).