Fix: Region pins not being movable after drawing the region
Something is not running as expected in the way that the raycaster was searching through all scene children and somehow the RegionLinesMesh was present over 20x and the hit boxes occupied a huge territory (I clicked around a lot and confirmed it), much larger than the lines that they draw.

So I reverted to a previous design in which I carefully control which meshes are selected for raycast intersection.

New problem (now that I can move the pins + redraw the region when doing so): moving the pins deletes the regionBoundaryMarker's "wrappedLongtitude" field. Ah. Right. Need to fix that.