Experiment: Failed to triangulate in 2D
I've finally become convinced (again) that this cannot be done in 2D. The poles make a complete wrap-around unavoidable. When I'm _not_ on the poles, I can perform wrap-around calculations, but that wrap-around won't save me when the longitude spans the entire longitude range and then wraps around to the beginning.

Revert to 3D.