Fix: Region movement snapping for one frame after starting to move the mouse
I didn't notice it until the region was shifted out from the globe a ways. Root cause: Not applying the offset quaternion to the first rotor. I dealt with it by applying the initial offset quaternion to the "mouse move" quaternion in the mouse handler rather than requiring everyone downstream to remember to do that. Made the calculation simpler.

Also updated "createWhere(...)" functions to only require the sphere radius (no use supplying all the region info when the calculation doesn't need it).