Fix: Extra line segments no longer appearing
I accidentally strayed onto a quirk of ThreeJs "lines": it uses OpenGL "line segments", which needs to be treated nicely. (1) You only need to specify one point at a time (that is, don't need both the beginning and the end after the first line segment, but only the end), and (2) a line will automatically be drawn from the end of the last segment to the new point (that is, if you are trying to draw disconnected lines in space, the line segment drawing will instead keep connecting point to point to point, which explains the unexpected web I had earlier).