Refactor: Moved geometry subdivision into a class, no more recursion
Rather than the series of functions that it used to be. I also took advantage of "delete array[index]" not actually eliminating the location from the array but instead just setting it to 0 to replace the recursion with a loop over an ever increasing array of "triangle indices" arrays.

Curiously, the line strip still works despite the (what I thought to be) disconnected triangles. Or maybe the triangles are not as disjointed as I thought?

Also have duplicate indices now in the lines, but I'm ok with that for the moment because it works.