Fix: Lines crossing each other, duplicate line segments
The package react-three/drei has a "Line" object that can draw line segments instead of a line strip, and it wasn't difficult to use (after I discovered that it existed).

Also fixed the duplication of line segments in the subdivision by using a dictionary, similar to the midpoint vertex generation.