Refactor: Not compiling yet, but making changes to how I'm storing data.
The current mental roadblock is good practice for what to do when I receive a DTO or JSON string (haven't figured out which to use yet) from the web interface and then figuring out which objects are new and which already existed. The big one is the image. Not sure how to tell if that is new or not without (a) exposing the PK to the GET request or (b) comparing values (expensive for a byte array) or (c) creating a an extra non-PK ID exclusively so that I can identify if the object is new or not when it returns from the website.