Refactor: Now creating the database again.
Had a problem with permissions. Figured it out and wrote a note about it. I needed to grant one of the SQL Server logins permissions on the database itself. There's a difference in being a server admin and a db_owner. A server admin lets you alter the server, while a db_owner lets you alter the database. They are intentionally separate.