Fix: Primary pin not moving, "lat,long" combo not displaying in the DetailsEdit"
The primary pin movement issue was due to setting the primary pin state in the PinMesh after the location had updated, but the EditableRegion reacted to that by reating a new pin with new bounding box and new new IDs. Fixed by making the EditableRegion only create a new pin when the primary pin's ID changed.

The detailsEdit just needed to be hooked up to the new state variables.