Fix: (1) Race condition in MouseHandler when ending click-and-drag. (2) Starting click-and-drag after clicking a mesh and quickly moving the mouse off the mesh before it kicked in wouldn't move the mesh.
Aside from the race condition (which took a good 2-3 hours to diagnose), I ended up doing a big refactor with the storing of intersections in the mouse state. This was necessary so that I could generalize the "initial offset quaternion" when enabling click-and-drag in MouseHandler without having to jump through some logic hoops just to figure out if the intersection under the cursor was the globe or something else. It ultimately didn't matter which mesh was below it, but my cursor intersection info in the mouse state didn't make that determination easy. Now it does. Feels like an obvious solution now that I should have come to earlier, but I didn't. :|