Experiment: Upgrading the "SingleAuthorEditMode" component with "edit" and "delete" buttons
And getting sidetracked for a while with the html "input" field making the element paralyzed to user input before managing to fish out of a google search that the input element's "value" field in React (I'm using Vite + React) needs to be bound to a state variable, but that didn't work, and finally a chance encounter with a stackoverflow post (https://stackoverflow.com/questions/34006333/cant-type-in-react-input-text-field) showed me the existence of the input's "defaultValue" field. That helped. The accepted answer said that the input's value can only be changed programmatically if you set the "value" property, not by typing. That explains things. The "defaultValue" property worked for what I needed.