Experiment: Still brainstorming the design, but I think I'm getting an idea of what I want from the top level (again).
There will be an array of sources (I think; still debating array vs json object). I will allow the import of a collection of sources. I will allow the creation of new sources. New sources will be appended to the array. The source state machine will keep track of all of these. That's the tricky part. There are repeated forms, and I want multiples to be editable at the same time (??is that asking too much of the design? will that be too hard to reconcile on the back end??).