Experiment: Trying to work out the basic UI structure of "add multiple sources, be able to edit all of them at the same time"
But I'm second-guessing whether the source material should be in a separate state slice or not. The stateSliceEditPoi has a lot going on in it already. My concern is the maintainability of all the support functions necessary to change all the details about the sources + all the details about the authors. But then I'll have to piece together different state machines, which could be worse...