Feature: Able to create multiple collapsible source entries and delete them
It feels a bit janky though. The javascript function makes a copy of any local variables that it uses at the time that the function is created (in this case, the sources). If I attempt to remove the latest entry, then the entry is not in the sources array because it hadn't been added at the time that its "deleteCallback" function had been created. But if I attempt to remove an older entry, then that _is_ in the sources array and is removed. Both result in the desired outcome, though the former by unintentional side effect.

Also learned that I must use an object to identify the sources in the state machine. The objects are stored as some kind of special "proxy" object and is difficult to dig into. But storing them in a javascript object with the temporary editId as key makes the proxy objects easy to tell apart.

Next: Start editing fields.