Experiment: Actually, I might not need that janky removal of source from the local EditEventSources "sources" array at the same time that I summoned the state machine to delete the same entry
The issue was that deleting that entry meant that there was (somehow, haven't figured it out yet) a change in state that caused the EventSource to re-render, but since the edit source had been deleted from the state machine, it came back null and that triggered several a lot of null reference errors. Solution: Return if the "editSource" retrieved from the state machine is null. And that was it. Now I can remove some of the jank.