Feature: Edit source publication lower bound year
I spent at least 2 hours trying to figure out input validation before eventually settling on "is the text a number". I tried to use input type number, but the input validation took me down a path of lots of extra boilerplate, so I opted to not do that and just use text.